pyproject.toml
[project]
name = "vox-scribe"
version = "0.1.0"
authors = [
  { name="Bob Ros", email="bobmeuchelm@gmail.com" },
]
license = "Apache-2.0"
description = "A real-time audio transcription tool/lib using Whisper."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pyaudio",
    "openai-whisper",
    "torch",
    "requests"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["vox"]

[project.urls]
Repository = "https://github.com/bob-ros2/vox"

[project.scripts]
vox = "vox.cli:main"
